fix gettimeofday_us return type to long
Reported by Yichun Zhang.

+++ runtime/kp_vm.c
@@ -49,7 +49,7 @@
 
 
 /* common helper function */
-int gettimeofday_us(void)
+long gettimeofday_us(void)
 {
 	struct timeval tv;
